backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic[email]==2.9.2
pydantic-settings==2.6.0
openai>=1.54.0
PyPDF2==3.0.1
python-docx==1.1.2
python-multipart==0.0.12
Jinja2==3.1.4
aiofiles==24.1.0
tenacity==9.0.0
PyJWT==2.9.0
supabase==2.9.1
pdf2image==1.17.0
